requirements.txt
Django==2.2.12
django-rest-framework==0.1.0
djangorestframework==3.11.0
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
